cdk/requirements.txt
aws-cdk.core==1.201.0
aws-cdk.aws-ec2==1.201.0
aws-cdk.aws-s3==1.201.0